move getClientPromptString
